electricity_market_api-main (1)/requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
chardet==5.1.0
click==8.1.3
et-xmlfile==1.1.0
fastapi==0.104.1
h11==0.14.0
idna==3.4
inflection==0.5.1
lxml==4.9.2
more-itertools==9.0.0
numpy==1.23.5
openpyxl==3.1.2
pandas==1.5.3
pydantic==2.5.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7
Quandl==3.7.0
requests==2.28.2
six==1.16.0
sentry-sdk==1.31.0
soupsieve==2.3.2.post1
starlette==0.27.0
toml==0.10.2
typing-extensions==4.8.0
uvicorn==0.22.0
wsproto==1.0.0
xlrd==2.0.1
xmltodict==0.13.0
excel-formulas-calculator==0.2.10
python-dotenv==1.0.1
ratelimit==2.2.1
stamina==24.2.0
tenacity==8.2.3
pysftp==0.2.9
cloudscraper==1.2.71
cachetools==5.5.2
paramiko<3